Initial take on a platform.sh environment repo for os2display
This commit adds the bulk of the tools and configuration needed for
running and operating an installation of os2display as a platform.sh
project.

The approach is based on the use of submodules and a multi-app approach
where each submodule is deployed as a separate platform.sh application.

The various os2display components requires some configuration during
deployment, and to aid that the environment repository comes with a
number of template files that are rendered and added to the applications
during build.

To make the administration of the submodules simpler this commit also
script that can automate the update of a submodule and the subsequent
commit. The script is called via a Taskfile.

The main README documents the basic use of the tools.

While the environment mostly works, there are still some final issues
with adding screens. This will most likely require som additional tweaks
to the setup.

NYOSD-20